requirements.txt
PyQt5==5.15.11
pillow==10.4.0
cryptography==43.0.1
